+++ src/neptune/types/type_casting.py
@@ -23,6 +23,7 @@ from typing import (
     Union,
 )
 
+from neptune.exceptions import UnsupportedType
 from neptune.internal.utils import (
     is_bool,
     is_dict_like,
@@ -84,8 +85,8 @@ def cast_value(value: Any) -> Value:
         return Namespace(value)
     elif (isinstance(value, list) and all(is_stringify_value(elem) for elem in value)) or from_stringify_value:
         return String(str(value))
-    else:
-        raise TypeError("Value of unsupported type {}".format(type(value)))
+
+    raise UnsupportedType(type_str=str(type(value)))
 
 
 def cast_value_for_extend(values: Union[Namespace, Series, Collection[Any]]) -> Union[Series, Namespace]:
@@ -114,5 +115,5 @@ def cast_value_for_extend(values: Union[Namespace, Series, Collection[Any]]) -> 
         return FloatSeries(values=values)
     elif from_stringify_value:
         return StringSeries(values=values)
-    else:
-        raise TypeError("Value of unsupported type List[{}]".format(type(sample_val)))
+
+    raise UnsupportedType(type_str=f"List[{type(sample_val)}]")