+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Removed `Run._short_id` property ([#1234](https://github.com/neptune-ai/neptune-client/pull/1234))
 - Removed `get_run_url` method ([#1238](https://github.com/neptune-ai/neptune-client/pull/1238))
 - Removed `neptune.new.sync` module ([#1240](https://github.com/neptune-ai/neptune-client/pull/1240))
+- Added exception for unsupported types ([#1229](https://github.com/neptune-ai/neptune-client/pull/1229))
 
 ## neptune-client 0.16.18